feat: add emojis and remove unused code
