+++ wondrous-app/components/CreateEntity/CreateEntityModal/TemplateBody/styles.tsx
@@ -1,16 +1,6 @@
-import { ButtonUnstyled, OptionUnstyled, PopperUnstyled, SelectUnstyled, TextareaAutosize } from '@mui/base';
-import SaveAltOutlined from '@mui/icons-material/SaveAltOutlined';
-import {
-  Autocomplete,
-  Grid,
-  Input,
-  InputAdornment,
-  InputBase,
-  MenuItem,
-  Select,
-  TextField,
-  Typography,
-} from '@mui/material';
+import { PopperUnstyled } from '@mui/base';
+import { Grid, Typography, Button } from '@mui/material';
+import EllipsesIcon from 'components/Icons/ellipsesIcon';
 
 import styled from 'styled-components';
 import palette from 'theme/palette';
@@ -20,6 +10,16 @@ export const StyledGrid = styled(Grid)`
     border: 1px solid ${palette.grey79};
     border-bottom-width: 0;
     padding: 12px;
+    max-height: 90vh;
+    overflow-y: scroll;
+    &::-webkit-scrollbar {
+      display: none;
+      width: 0;
+      height: 0;
+    }
+    /* Hide scrollbar for IE, Edge and Firefox */
+    -ms-overflow-style: none; /* IE and Edge */
+    scrollbar-width: none; /* Firefox */
     &:nth-child(even) {
       border-right-width: 0;
       border-left-width: 0;
@@ -138,4 +138,71 @@ export const SaveTemplateButtonText = styled(Typography)`
 
 export const TemplateSelectText = styled.span`
   text-transform: capitalize;
+`;
+
+export const TaskTemplateEllipsesIcon = styled(EllipsesIcon)`
+  width: 24px;
+  color: ${palette.white};
+  background: rgba(122, 122, 122, 0.2);
+  height: 24px;
+  padding: 6px;
+  border-radius: 6px;
+  path {
+    background: rgba(122, 122, 122, 0.2);
+  }
+  :hover {
+    cursor: pointer;
+  }
+`;
+
+export const TaskTemplateOptionsDeleteLabel = styled(Button)`
+  && {
+    font-family: 'Space Grotesk';
+    font-size: 13px;
+    font-weight: 100;
+    width: 100%;
+    padding: 4px 8px 4px 8px;
+    color: #cf4141;
+    margin-bottom: 4px;
+    overflow: hidden;
+    border-radius: 4px;
+    background: #121212;
+    :hover {
+      background: rgba(122, 122, 122, 0.2);
+    }
+  }
+`;
+
+export const TaskTemplateOptionsLabel = styled(Button)`
+  && {
+    font-family: 'Space Grotesk';
+    font-size: 13px;
+    font-weight: 100;
+    padding: 4px 8px 4px 8px;
+    color: ${palette.white};
+    width: 100%;
+    margin-bottom: 4px;
+    overflow: hidden;
+    border-radius: 4px;
+    background: #121212;
+    :hover {
+      background: rgba(122, 122, 122, 0.2);
+    }
+  }
+`;
+
+export const TemplateOptionsPopper = styled(PopperUnstyled)`
+  padding: 4px;
+  border-radius: 4px;
+  background-color: #1f1f1f;
+  border: 1px solid ${palette.grey57};
+  position: relative;
+  z-index: 9999;
+`;
+
+export const StyledGridContainer = styled(Grid)`
+  && {
+    max-height: 100vh;
+    overflow-y: scroll;
+  }
 `;
